Added: Support for recent photos and albums
Could be buggy, it does return some recent stuff, but doesnt match exactly
with the recent photos and albums returned actually on facebook (there
are some matches). Need to check if I am sorting by the
right field (and if that field is actually sortable with the right meaning).

But commiting it for fun!